+++ interop/src/commonMain/kotlin/io/github/durun/vst3kotlin/vst/AttrProperty.kt
